Fix issue with removing ./out folder

+++ deploy/powershell/Publish-Site.ps1
@@ -11,7 +11,10 @@ Push-Location $(./Join-Path-Recursively.ps1 -pathParts "..,..,ui".Split(","))
 Write-Host "===========================================================" -ForegroundColor Yellow
 Write-Host " Building the website" -ForegroundColor Yellow
 Write-Host "===========================================================" -ForegroundColor Yellow
-Remove-Item -Path ./out -Recurse -Force
+if (Test-Path ./out)
+{
+   Remove-Item -Path ./out -Recurse -Force
+}
 Start-Sleep -Seconds 10
 npm ci
 npm run build
